BCDA-4826: Casting

+++ bcdaworker/worker/worker.go
@@ -66,7 +66,7 @@ func (w *worker) ProcessJob(ctx context.Context, job models.Job, jobArgs models.
 	defer t.Close()
 	ctx = metrics.NewContext(ctx, t)
 	ctx, c := metrics.NewParent(ctx, fmt.Sprintf("ProcessJob-%s", jobArgs.ResourceType))
-	metrics.AddAttribute(ctx, "jobId", strconv.FormatUint(job.ID), 10)
+	metrics.AddAttribute(ctx, "jobId", strconv.FormatUint(uint64(job.ID)), 10)
 	metrics.AddAttribute(ctx, "acoId", jobArgs.ACOID)
 	metrics.AddAttribute(ctx, "resourceType", jobArgs.ResourceType)
 	metrics.AddAttribute(ctx, "requestUrl", job.RequestURL)
